Create calculus repositories and test

+++ hdv_watcher/lib/features/item/data/model/item_model.dart
@@ -35,7 +35,7 @@ class ItemModel extends Item {
       name: json["name"],
       imgUrl: json["img_url"],
       ressourceType: ressourceFromString(name: ressourceTypeName),
-      petXp: json["f_xp"] ?? 0,
+      petXp: json["f_xp"].toDouble() ?? 0.0,
       quantityForUnitXp: json["xp_quantity"] ?? 0,
       superPrice: SuperPrice.fromItemFactory(
           unitPrices: unitPrices,
